chore: extract user-specific helpers
To keep those in `users` module.
